feat(media-transcription): rename files
