+++ src/features/media-transcription/mediaTranscription.service.ts
@@ -1,21 +1,14 @@
 import { IBalancedSegment } from '@/types/youtube/youtube.type';
-import { readFile, unlink, writeFile } from 'fs/promises';
-import { UPLOADS_DIR_PATH } from '../../whisper/local/whisper.constant';
 import logger from '@/utils/logger';
-import path from 'path';
 import type { Express } from 'express';
-import { v4 as uuidv4 } from 'uuid';
 import { InternalServerError } from '@/core/error';
+import whisperService from '../whisper/api/whisper.service';
 import transcriptService from '@/services/transcript/transcript.service';
-import whisperService from '../../whisper/local/whisper.service';
 
-class AudioTranscriptionService {
-    private async _getTranscriptSegmentsFromAudio(audioFile: Express.Multer.File) {
-        const tempInputAudioPath = await this.saveAudioTempFile(audioFile);
-        let tempOutputPath = null;
+class MediaTranscriptionService {
+    public async getTranscriptSegmentsFromFile(file: Express.Multer.File) {
         try {
-            tempOutputPath = await whisperService.runWhisperLocal(tempInputAudioPath);
-            const rawTranscript = await readFile(tempOutputPath, 'utf-8');
+            const rawTranscript = await whisperService.getTranscriptionFromFile(file);
             const parseResult = this.getTranscriptSegments(rawTranscript);
             if (!parseResult.ok) {
                 throw new Error(parseResult.error);
@@ -28,23 +21,8 @@ class AudioTranscriptionService {
             return balancedSegments;
         } catch (err) {
             logger.error('Failed to process transcript', err);
-            throw new InternalServerError('Failed to get transcription from audio.');
-        } finally {
-            this.safeUnlink(tempInputAudioPath);
-            this.safeUnlink(tempOutputPath);
+            throw new InternalServerError('Failed to get transcription.');
         }
-    }
-
-    private async saveAudioTempFile(file: Express.Multer.File) {
-        const ext = path.extname(file.originalname).toLowerCase();
-        const pathToSave = path.join(UPLOADS_DIR_PATH, `${uuidv4()}${ext}`);
-        await writeFile(pathToSave, file.buffer);
-        return pathToSave;
-    }
-
-    private async safeUnlink(path: string | null) {
-        if (!path) return;
-        await unlink(path).catch(() => {});
     }
 
     private getTranscriptSegments(
@@ -75,13 +53,11 @@ class AudioTranscriptionService {
     }
 
     private getSeconds(timestamp: string) {
-        const firstRow = timestamp.split(',')[0];
-        const secondRowsSplited = firstRow.split(':');
-        const hours = Number(secondRowsSplited[0]);
-        const minutes = Number(secondRowsSplited[1]);
-        const seconds = Number(secondRowsSplited[2]);
-        return hours * 3600 + minutes * 60 + seconds;
+        const [timePart, msPart] = timestamp.split(',');
+        const [hours, minutes, seconds] = timePart.split(':').map(Number);
+        const ms = msPart ? Number(msPart) / 1000 : 0;
+        return hours * 3600 + minutes * 60 + seconds + ms;
     }
 }
 
-export default new AudioTranscriptionService();
+export default new MediaTranscriptionService();
